Deleted all testing uploads, created a new s3 file
with a set up to update the pictures in the future. Update only works
for text fields right now. Path is not translating to a URL, String
or buffer when uploading so I changed PATCH back to its original state
for now I am hiding the picture upload function from the update form
in the front end, user can create but in order to get rid of the picture
they must delete it. I am not satisfied with this implementation and
will work on it in the future. Because of that, I changed the file
upload from required to not required on the schema. I also created an
array of skills.